requirements.txt
aircraft-auralization==1.0.0
appdirs==1.4.4
audioread==3.0.0
certifi==2022.12.7
cffi==1.15.1
cftime==1.6.2
charset-normalizer==3.0.1
coloredlogs==15.0.1
colorednoise==2.1.0
contourpy==1.0.7
cycler==0.11.0
DateTime==5.0
decorator==4.4.2
et-xmlfile==1.1.0
fonttools==4.38.0
humanfriendly==10.0
idna==3.4
imageio==2.25.0
imageio-ffmpeg==0.4.8
joblib==1.2.0
kiwisolver==1.4.4
librosa==0.9.2
llvmlite==0.39.1
matplotlib==3.6.3
moviepy==1.0.3
netCDF4==1.6.2
numba==0.56.4
numpy==1.23.5
opencv-python==4.7.0.68
openpyxl==3.1.2
packaging==23.0
pandas==2.0.1
Pillow==9.4.0
pooch==1.6.0
proglog==0.1.10
pycparser==2.21
pyfilterbank @ git+https://github.com/SiggiGue/pyfilterbank.git@87b9bb52cb6ecdbcbc0385f1c10ed040b22b49d1
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
python-dateutil==2.8.2
python-sofa==0.2.0
pytz==2022.7.1
pyvista==0.37.0
requests==2.28.2
resampy==0.4.2
samplerate==0.1.0
scikit-learn==1.2.1
scipy==1.10.0
scooby==0.7.1
shapely==2.0.1
six==1.16.0
sounddevice==0.4.5
soundfile==0.11.0
threadpoolctl==3.1.0
tqdm==4.64.1
tzdata==2023.3
ui==0.1.4
urllib3==1.26.14
vtk==9.2.5
yulewalker==0.1.1
zope.interface==5.5.2
